refactor(settings): post-commit cleanup for Tap to Walk feature
- Use EARTH_RADIUS_METERS constant and kotlin.math.PI in TapToWalkOverlay.computeWalkTarget
- Guard onDismissed() in TapToWalkOverlay.dismiss() — only fires when overlay was actually shown
- Replace tapToWalkScaleMpxFlow MutableStateFlow middleman with stateIn in FloatingWidgetService
- Add smoke tests: Tap to Walk section, Location Memory, Favorites, Routes sub-screen headings
- Split SettingsScreen.kt (1528 lines) into SettingsGpsSubScreen, SettingsMenusSubScreen, SettingsFavoritesRoutesSubScreen
- Change tapToWalkScaleMpx model type from Double to Int throughout (AppConstants, DataStore key renamed to tap_to_walk_scale_mpx_int, action, state, VM, repository, tests)

+++ core/data/src/main/kotlin/com/locationjoystick/core/data/SettingsRepository.kt
@@ -214,7 +214,7 @@ class SettingsRepository
 
         fun getTapToWalkOverlayEnabled(): Flow<Boolean> = dataSource.getTapToWalkOverlayEnabled()
 
-        fun getTapToWalkScaleMpx(): Flow<Double> = dataSource.getTapToWalkScaleMpx()
+        fun getTapToWalkScaleMpx(): Flow<Int> = dataSource.getTapToWalkScaleMpx()
 
         fun getSettingsSnapshot(): Flow<SettingsSnapshot> = dataSource.getSettingsSnapshot()
