+++ feature/settings/impl/src/main/kotlin/com/locationjoystick/feature/settings/impl/SettingsViewModel.kt
@@ -120,7 +120,7 @@ class SettingsViewModel
             val selectedHotRouteIds: Set<String>? = null,
             val floatingMapQuickWalk: Boolean? = null,
             val tapToWalkOverlayEnabled: Boolean? = null,
-            val tapToWalkScaleMpx: Double? = null,
+            val tapToWalkScaleMpx: Int? = null,
         )
 
         private val mutableDraft = MutableStateFlow(DraftState())
@@ -330,7 +330,7 @@ class SettingsViewModel
             mutableDraft.update { it.copy(tapToWalkOverlayEnabled = enabled) }
         }
 
-        fun setTapToWalkScaleMpx(scale: Double) {
+        fun setTapToWalkScaleMpx(scale: Int) {
             mutableDraft.update { it.copy(tapToWalkScaleMpx = scale) }
         }
