+++ feature/settings/impl/src/test/kotlin/com/locationjoystick/feature/settings/impl/SettingsViewModelSaveTest.kt
@@ -559,5 +559,5 @@ internal class SaveTestPreferencesDataSource : PreferencesDataSource {
 
     override fun getTapToWalkOverlayEnabled(): Flow<Boolean> = flowOf(false)
 
-    override fun getTapToWalkScaleMpx(): Flow<Double> = flowOf(AppConstants.TapToWalkConstants.DEFAULT_SCALE_MPX)
+    override fun getTapToWalkScaleMpx(): Flow<Int> = flowOf(AppConstants.TapToWalkConstants.DEFAULT_SCALE_MPX)
 }
